fix: read the status visibility setting through the settings service
