+++ apps/meteor/server/api/v1/im.ts
@@ -26,7 +26,8 @@ import { hasPermissionAsync } from '../../lib/authorization/hasPermission';
 import { eraseRoom } from '../../lib/eraseRoom';
 import { openRoom } from '../../lib/openRoom';
 import { getRoomByNameOrIdWithOptionToJoin } from '../../lib/rooms/getRoomByNameOrIdWithOptionToJoin';
-import { canSeeStatus, getHiddenFrom, redactStatus } from '../../lib/statusVisibility/canSeeStatus';
+import { canSeeStatus, getHiddenFrom } from '../../lib/statusVisibility/canSeeStatus';
+import { redactStatus } from '../../lib/statusVisibility/redactStatus';
 import { blockUserMethod } from '../../lib/users/blockUser';
 import { unblockUserMethod } from '../../lib/users/unblockUser';
 import { normalizeMessagesForUser } from '../../lib/utils/lib/normalizeMessagesForUser';
