+++ apps/meteor/server/api/v1/users.ts
@@ -48,7 +48,9 @@ import { SystemLogger } from '../../lib/logger/system';
 import { notifyOnUserChange, notifyOnUserChangeAsync } from '../../lib/notifyListener';
 import { resetUserE2EEncriptionKey } from '../../lib/resetUserE2EKey';
 import { validateNameChars } from '../../lib/shared/validateNameChars';
-import { canSeeStatus, getHiddenFrom, redactStatus, resolveUsersByIds } from '../../lib/statusVisibility/canSeeStatus';
+import { canSeeStatus, getHiddenFrom } from '../../lib/statusVisibility/canSeeStatus';
+import { redactStatus } from '../../lib/statusVisibility/redactStatus';
+import { resolveUsersByIds } from '../../lib/statusVisibility/resolveUsers';
 import { checkEmailAvailability } from '../../lib/users/checkEmailAvailability';
 import { checkUsernameAvailability, checkUsernameAvailabilityWithValidation } from '../../lib/users/checkUsernameAvailability';
 import { deleteUser } from '../../lib/users/deleteUser';
